+++ backend/MerkERP.Core/Models/ItemType_s.cs
@@ -2,7 +2,7 @@ namespace MerkERP.Core.Models;
 
 public class ItemType_s
 {
-    public int ItemTypeId { get; set; }
+    public long ItemTypeId { get; set; }
     public string Name { get; set; } = string.Empty;
 
     public ICollection<Item_cs> Items { get; set; } = [];
