+++ backend/MerkERP.Core/Models/WareHouse_cs.cs
@@ -2,9 +2,9 @@ namespace MerkERP.Core.Models;
 
 public class WareHouse_cs
 {
-	public int WarehouseId { get; set; }
+	public long WarehouseId { get; set; }
 	public string Name { get; set; } = string.Empty;
-	public int? ParentWarehouseId { get; set; }
+	public long? ParentWarehouseId { get; set; }
 	public bool IsGroup { get; set; }
 	public bool IsActive { get; set; } = true;
